dynamic camera settings
i've it set up so you pick between 10 levels of how fast it goes, also to pull up the pause menu you press the P key for now! (also the leave game button does nothing until you build the game! and the continue does what you'd expect it to!) the pause menu does not pause the game, this is intentional so that when you try to leave and being attacked you will still be attacked!