a cakewalk song
just for testing purposes, also do ask before using a song, whether on my youtube or scratch with a link to where and how you want to use it. any one who avoids this (same for the msc songs) will be taken down unless it's gameplay footage, but that will also be only limited to full builds, and if not stressed enough, you cannot go about and post gnty in any state anywhere claiming it's yours. this github will act as proof. (just clearing up this for any legal disputes over my songs and game! it is open source, but doesnt mean its completely free to use, like the characters (like liz) were not designed by me and i do not own them, and simply have permission to use them!)