they now sing in the testing scene
it's about time i start the devlog videos, so here is the first one to get recorded for it! (mainly going over the models and the spatial based sound and the settings menu so far!) i will go moe depth when i actually start script writing, also i do plan on redoing baningo and liz's models though so don't get attatched to them!