bug fixing
just a minor tweak, originally (as this is trigger based to pick an item up) if you held an object you can collect it again while it's in your hand and have a duplicate, this is not intended, but with this script adjustment (and yes you'll be able to hold weapons soon) that problem is fixed! when set up properly now it'll look how it did before but now it will also allow for animations while it's in your hand as well! (maybe you'll pick up a creature or animated peice and it could actually move in your hand!) :P