drop system
you can now drop items your holding, of course though that they have no physics attatched they float, but that is a simple fix! now the inventory system is complete! (too bad that every script rellies on eachother though, i wish i made it simpler but as it works, i see no reason in doing anything different, and the interface is simple and straight forward at least and so is the set up!)