thw droppening
now when you drop the whillow doll (or walk over it) it'll move either down, or even roll!  this was just the addition of a rigidbody and a capsule collider, though i do want to add doll physics soon! also i adjusted the code that i just pushed so that what i couldnt figure out wont send hundreds of error messages!