cutscenes ingame can now be done
for testing purposes i have the NON CANONICAL warped version of chikn make an apearance to test it. I did use inspiration to how it's done in finding frankie with this being that you can still freely look around and move, but a barrier is put around the player so they cannot move if needed for the moment, i also made it a bool to be optional. so i could also assign a direct chase sequence (part or whole) to it and you wont be stuck and can run. dont know yet where to go from here as so much else needs done, though i might make the object placement next.