import done for whillow
just assembled the puppet for whillow, i'm going to start work now on object interaction between things. (i plan on having it so you can give whillow the doll of herself to test this!)

the eye rig for whillow is broken. i'm currently trying to think of a way to not have this problem, though i could always do the inching method when animating to get it how i want it!