whillow's base position
not much to say, but that i'm doubling down onto the power idea for the creatures, and whillow's is being stretchy. this is exagerated in the cartoon design but for the normal one it is just like a spring if that is a good visual! (some creatures have no powers or special abilities, they instead have very specific hobbies or act as a protaganist who needs not an outside force to help them!)