the negative fifth floor
i've placed the 3 creatures that are in this area that i've modeled so far into their places. a temprary texture has been applied to the floor's colors, and there is still a nedd for -4 and -3 before i can start assembling more of the game. (real quick to the only person who came ths far, -4 and -3 will be seperated into areas instead of one big area that gets loaded in. the same will go for -5 when more comes, but do know that the floors will be one giant model that is loaded all at once, then will turn off and turn on as needed by the game depending on where you are. this is just for performance reasons that way slower appliances can run the game faster. though to open up the entire game from the eitor will only be resserved for what i geuss would need to be a upper end computer.)   One quick thing, the walls will have a much bigger texture then the creatures as well due to it's size and for ease on my part it will go from carpet and hardwood to just a mysterious substance for flooring as going about and individually coloring all of it would be a nightmare. :L   Though the walls will be given unique textures and normal maps so it matches the chracters, and again, n a much bigger file. maybe double or quadruple the size it is for the creatures themselves.   ALso quick note to self here; generation-> solidify modifier->thickness=0.0001

the walls will all be one color though, but more detailing will be done! (mainly water damage and wear/tear overall mostly and this does allow me to make more complex ones if i wish as well though! like if i wanted i can have a room have claw marks and such, though that'd need to be its own  hover mesh ingame though, but for simple doublesideable details it will be water damage and wear/tear!)