area 1 of floor -4 modeled
not much to it, its just the area that crusty, liz, and rebec reside. though i may change how you go between floors from a stairwell to an elevator. though do know that that doesnt mean chapter 2 is  floor -4 and that there will only be 5 chapters, there will actually be 8-10 chapters and before that there are 8 floors to explore in totality! and also it would be a good time to explain now that the GNty facility is absolutely massive, though in chapter 1 most of your time is spent in floor -4! (also there is only one area for floor -5 as that is all that is planned for chapter1, though do know that it will be exteded for an upcoming chapter!)