area 2 of floor -4
not much to it, just a quick area, though there are 3 more for floor -4, then a couple for floor -3 left. then i'm going to start 3d modeling again for the missing chraacters once the map is done. i may add a staircase model though to get to floor -3 but that will be it for a while aside from chracters. and yes i will add their sounds to the testing midi, just know they kinda change the mood that the song currently has, and a couple go more into bodily horror!