shivr#1
no rig or pupils yet. but next update will include them!
(as i complete the map i'll be adding the creatures to their areas. next up is a more residential are with only 3 creatures though.) not much to say but i dont really like how shivr's eyes are popped out but it looks good though its just it looks weird.